refactor: Input을 SearchInput으로 변경
